+++ salim/crawlers/providers/yohananof.py
@@ -12,9 +12,9 @@ urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "../../..")))
 
-from app.crawlers.base import CrawlerBase
-from app.crawlers.utils.html_utils import extract_file_links
-from app.crawlers.utils.file_utils import (create_provider_dir, extract_file_info, download_file_with_session, transfer_cookies)
+from crawlers.base import CrawlerBase
+from crawlers.utils.html_utils import extract_file_links
+from crawlers.utils.file_utils import (create_provider_dir, extract_file_info, download_file_with_session, transfer_cookies)
 
 LOGIN_URL = "https://url.publishedprices.co.il/login"
 PROVIDER_URL = "https://url.publishedprices.co.il/file"
